+++ CoreDataAccessView/src/au/gov/asd/tac/constellation/views/dataaccess/plugins/experimental/TestParametersPlugin.java
@@ -387,7 +387,8 @@ public class TestParametersPlugin extends RecordStoreQueryPlugin implements Data
         // Local process-tracking variables (Process is indeteminate due to the nature of plugin reporting through the logger)
         final int currentProcessStep = 0;
         final int totalProcessSteps = -1;
-        interaction.setProgress(currentProcessStep, totalProcessSteps, "Testing parameters...", true);
+        interaction.setProgressTimestamp(true);
+        interaction.setProgress(currentProcessStep, totalProcessSteps, "Testing parameters...", true, parameters);
 
         //Display parameter information
         LOGGER.log(Level.INFO, "parameters: {0}", parameters);