+++ CoreDataAccessView/src/au/gov/asd/tac/constellation/views/dataaccess/plugins/importing/ExtractTypesFromTextPlugin.java
@@ -103,8 +103,9 @@ public class ExtractTypesFromTextPlugin extends RecordStoreQueryPlugin implement
         
         // Local process-tracking varables (Process is indeteminate until quantity of entities is known)
         int currentProcessStep = 0;
-        int totalProcessSteps = -1; 
-        interaction.setProgress(currentProcessStep, totalProcessSteps, "Extracting entities...", true);
+        int totalProcessSteps = -1;
+        interaction.setProgressTimestamp(true);
+        interaction.setProgress(currentProcessStep, totalProcessSteps, "Extracting entities...", true, parameters);
  
         // Determine number of entity types 
         final List<ExtractedVertexType> extractedTypes = SchemaVertexTypeUtilities.extractVertexTypes(text);