+++ CorePluginFramework/src/au/gov/asd/tac/constellation/plugins/text/TextPluginInteraction.java
@@ -17,7 +17,10 @@ package au.gov.asd.tac.constellation.plugins.text;
 
 import au.gov.asd.tac.constellation.plugins.PluginInteraction;
 import au.gov.asd.tac.constellation.plugins.PluginNotificationLevel;
+import au.gov.asd.tac.constellation.plugins.parameters.PluginParameter;
 import au.gov.asd.tac.constellation.plugins.parameters.PluginParameters;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -62,6 +65,13 @@ public class TextPluginInteraction implements PluginInteraction {
     }
     
     @Override
+    public void setProgress(final int currentStep, final int totalSteps, final String message, final boolean cancellable, final PluginParameters parameters) throws InterruptedException {
+        currentMessage = message;
+        final Map<String, PluginParameter<?>> params = parameters != null ? parameters.getParameters() : new HashMap<>();
+        LOGGER.log(Level.INFO, "currentStep={0} totalSteps={1} message={2} parameters={3}", new Object[]{currentStep, totalSteps, message, params.toString()});
+    }
+    
+    @Override
     public void setProgress(final int currentStep, final int totalSteps, final boolean cancellable) throws InterruptedException {
         LOGGER.log(Level.INFO, "currentStep={0} totalSteps={1}", new Object[]{currentStep, totalSteps});
     }
@@ -90,6 +100,18 @@ public class TextPluginInteraction implements PluginInteraction {
     @Override
     public boolean prompt(final String promptName, final PluginParameters parameters, final String disclaimer, final String helpID) {
         throw new UnsupportedOperationException(NOT_SUPPORTED); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    @Override
+    public void setProgress(final int currentStep, final int totalSteps, final String message, final boolean cancellable, final PluginParameters parameters, final int selected) throws InterruptedException {
+        currentMessage = message;
+        final Map<String, PluginParameter<?>> params = parameters != null ? parameters.getParameters() : new HashMap<>();
+        LOGGER.log(Level.INFO, "currentStep={0} totalSteps={1} message={2} parameters={3} selected={4}", new Object[]{currentStep, totalSteps, message, params.toString(), selected});
+    }
+
+    @Override
+    public void setProgressTimestamp(boolean addTimestamp) throws InterruptedException {
+        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }
 
 }